fix: resolve arch violations from SessionTelemetry refactor
- Extract git LOC helpers to _headless_git.py to keep headless.py under 750-line budget
- Move _build_session_telemetry to _headless_result.py (correct module boundary)
- Remove unused _NO_TELEMETRY module-level constant (eliminated singleton guard violation)
- Fix session_log.py to import SessionTelemetry from public autoskillit.core gateway
- Add test_session_log_flush.py pipeline-layer exception to test_layer_enforcement.py
- Update _LEGACY_JSON_WRITES line numbers to match refactored session_log.py
- Update execution/ exemption to 37 files and document _headless_git.py in CLAUDE.md
