+++ tests/arch/test_layer_enforcement.py
@@ -1172,6 +1172,7 @@ _TEST_LAYER_ALLOWLIST: dict[str, frozenset[str]] = {
     "tests/execution/test_clone_guard.py": frozenset({"autoskillit.pipeline"}),
     "tests/execution/test_commands.py": frozenset({"autoskillit.cli"}),
     "tests/execution/test_headless_core.py": frozenset({"autoskillit.pipeline"}),
+    "tests/execution/test_session_log_flush.py": frozenset({"autoskillit.pipeline"}),
     "tests/execution/test_headless_path_validation.py": frozenset(
         {"autoskillit.pipeline", "autoskillit.recipe"}
     ),
