+++ tests/arch/test_subpackage_isolation.py
@@ -703,7 +703,10 @@ def test_no_subpackage_exceeds_10_files() -> None:
         _merge_queue_group_ci.py extracts merge-group CI helpers and GraphQL mutation/query
         strings from merge_queue.py to satisfy the 500-line size budget (P8-F4 follow-up),
         bringing the count to 36.
-        Exempt at 36 files.
+        _headless_git.py extracts git LOC-capture helpers (_capture_git_head_sha,
+        _parse_numstat, _compute_loc_changed) from headless.py to keep it under the
+        750-line architectural budget, bringing the count to 37.
+        Exempt at 37 files.
       core/ — REQ-CNST-003-E4: core/ types split into per-concern type modules
         (_type_enums, _type_protocols_logging, _type_protocols_execution,
         _type_protocols_github, _type_protocols_workspace, _type_protocols_recipe,
@@ -776,7 +779,7 @@ def test_no_subpackage_exceeds_10_files() -> None:
     EXEMPTIONS: dict[str, int] = {
         "server": 25,
         "recipe": 50,
-        "execution": 36,
+        "execution": 37,
         "core": 33,
         "cli": 42,
         "hooks": 28,
